feat: add moves table in Pokemon Detail (w/ TODO)
This involved modifying TableRow and TableCell components to use an extra
supplied class name.

A separate MovesApi layer was also made to perform api calls.

Also, an entire separate component called MovesTable was made to render
all the moves learned by the Pokemon.

One issue is that all the table headers are aligned to the right. They
need to be aligned to the left.