feat: add static images, extend TableCellHeader
images for the rarity circles and damage class icons were added.
Separate folders were made for damage class icons and the rarity images.

Also, an extra prop was added to TableCellHeader so that it can be used
in actual tabular data and semi-tabular data (like in Pokedex data,
training info, breeding info and so on).