fix(item): use item ids for fetching instead of names in listing page, don't fetch gen 8+ items
This was done because an item having an apostrophe symbol in its name caused issues when fetching
its data. Moreover, the data for gen 8+ items was also being fetched, and this has also been fixed
by identifying the cut-off item id and filtering items that have ids above this value.